requirements.txt
defusedxml==0.4.1
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.9.1
django-allauth==0.24.1
django-crispy-forms==1.5.2
django-debug-toolbar==1.4
django-embed-video==1.1.0
django-toolbelt==0.0.1
fake-factory==0.5.3
gunicorn==19.4.5
oauthlib==1.0.3
Pillow==3.1.0
psycopg2==2.6.1
python3-openid==3.0.9
requests==2.9.1
requests-oauthlib==0.6.1
sqlparse==0.1.19
static3==0.6.1
